added dapper handler, added filter extension

+++ backend/src/AnimalAllies.Application/Models/PagedList.cs
@@ -5,7 +5,7 @@ public class PagedList<T>
     public IReadOnlyList<T> Items { get; init; } = [];
     public int PageSize { get; init; }
     public int Page { get; init; }
-    public int TotalCount { get; init; }
+    public long TotalCount { get; init; }
     public bool HasNextPage => Page * PageSize < TotalCount;
     public bool HasPreviousPage => Page * PageSize > 1;
 }
